refactor(config): refactor config
