+++ bin/proxy/config.js
@@ -16,36 +16,15 @@ const defaultValue = plug('default/config.default.js');
 const processArgs = plug('util/process.args.js');
 const cwd = process.cwd();
 const currConfig = path.join(cwd, 'tsw.config.js');
-let isFirstLoad = false;
 let cache = {
     config: null
 };
-
 
 if (global[__filename]) {
     cache = global[__filename];
 } else {
     global[__filename] = cache;
-    isFirstLoad = true;
 }
-
-if (isFirstLoad) {
-    process.dlopen = (function(fn) {
-        const parent = path.join(__dirname, '..');
-
-        return function(module, curr) {
-            // 检查node私有文件
-            if (/\.node$/i.test(curr) && curr.indexOf(parent) !== 0) {
-                // 发现私有node扩展
-                setTimeout(function() {
-                    require('runtime/md5.check.js').findNodeCpp(curr);
-                }, 3000);
-            }
-            return fn.apply(this, arguments);
-        };
-    })(process.dlopen);
-}
-
 
 if (typeof processArgs.config === 'string') {
     cache.config = path.join(cwd, processArgs.config);
