+++ bin/proxy/index.js
@@ -11,6 +11,7 @@
 const plug = require('../tsw/plug.js');
 
 require('./version.js');
+require('./addons.finder.js');
 
 plug('runtime/Console.hack.js');
 plug('runtime/fs.hack.js');
